Add get limit by category id

+++ android/app/src/main/java/io/github/kn65op/domag/database/entities/CategoryLimit.kt
@@ -7,7 +7,7 @@ import io.github.kn65op.android.lib.type.FixedPointNumber
 @Entity
 data class CategoryLimit(
     @PrimaryKey
-    val uid: Int? = null,
+    val uid: Long? = null,
 
     val categoryId: Int,
